Consolidate block body logic
